Created tunnel_comm (separated ping & status)
