+++ Makefile
@@ -196,7 +196,7 @@ PROJ_OBJ += activeMarkerDeck.o
 # Tunnel exploration
 PROJ_OBJ += tunnel.o tunnel_config.o tunnel_ping.o tunnel_parameters.o 
 PROJ_OBJ += tunnel_avoider.o tunnel_behavior.o tunnel_commander.o
-PROJ_OBJ += tunnel_signal.o tunnel_relay.o
+PROJ_OBJ += tunnel_signal.o tunnel_relay.o tunnel_comm.o
 
 ifeq ($(LPS_TDOA_ENABLE), 1)
 CFLAGS += -DLPS_TDOA_ENABLE
