+++ src/modules/src/tunnel.c
@@ -12,12 +12,9 @@
 #include "tunnel.h"
 
 #include "tunnel_config.h"
-#include "tunnel_ping.h"
 #include "tunnel_commander.h"
-#include "tunnel_parameters.h"
 #include "tunnel_behavior.h"
-#include "tunnel_signal.h"
-#include "tunnel_relay.h"
+#include "tunnel_comm.h"
 
 #define DEBUG_MODULE "TUN"
 #include "debug.h"
@@ -58,76 +55,36 @@ static DroneRole droneRole;
 DroneRole tunnelGetDroneRole() { return droneRole; }
 void tunnelSetDroneRole(DroneRole newRole) { droneRole = newRole; }
 
-typedef enum {
-  CRTP_TUNNEL_CHANNEL_PING      = 0x00,
-  CRTP_TUNNEL_CHANNEL_PARAM     = 0x01,
-  CRTP_TUNNEL_CHANNEL_COMMANDER = 0x02,
-  CRTP_TUNNEL_CHANNEL_COMMAND   = 0x03,
-} CRTPTunnelChannel;
-
-typedef enum {
-  CRTP_TUNNEL_COMMAND_TAKE_OFF = 0x00, // Start flying and do our thing!
-  CRTP_TUNNEL_COMMAND_LAND     = 0x01, // Land no matter where we are (e.g. to manually save battery)
-  CRTP_TUNNEL_COMMAND_RTH      = 0x02, // Return to home automatically
-  CRTP_TUNNEL_COMMAND_STOP     = 0x03, // Stop the motors and return to Idle state (for emergencies or tests)
-} CRTPTunnelCommand;
-
 static void tunnelTask(void *param) {
-    systemWaitStart();
-    vTaskDelay(2000);
+  systemWaitStart();
+  vTaskDelay(2000);
 
-    // Set the max radio TX Power
-    radiolinkSetPowerDbm(4);
+  // Set the max radio TX Power
+  radiolinkSetPowerDbm(4);
 
-    // Stop red LED (it gets annoying ^^)
-    ledseqStop(SYS_LED, seq_alive);
-    ledseqStop(SYS_LED, seq_calibrated);
-    ledSet(SYS_LED, false);
+  // Stop red LED (it gets annoying ^^)
+  ledseqStop(SYS_LED, seq_alive);
+  ledseqStop(SYS_LED, seq_calibrated);
+  ledSet(SYS_LED, false);
 
-    // Print some information
-    DEBUG_PRINT("Drone ID: %i\n", getDroneId());
+  // Print some information
+  DEBUG_PRINT("Drone ID: %i\n", getDroneId());
 
-    TickType_t lastWakeTime = xTaskGetTickCount();
+  TickType_t lastWakeTime = xTaskGetTickCount();
 
-    while (1) {
-      vTaskDelayUntil(&lastWakeTime, M2T(1000 / TUNNEL_TASK_RATE_HZ));
+  while (1) {
+    vTaskDelayUntil(&lastWakeTime, M2T(1000 / TUNNEL_TASK_RATE_HZ));
 
-      // Calculate the new drone movement and send it to the stabilizer
-      tunnelCommanderUpdate();
+    // Calculate the new drone movement and send it to the stabilizer
+    tunnelCommanderUpdate();
 
-      // Handle ping routines
-      tunnelPingUpdate();
-    }
-}
-
-void crtpTunnelHandler(CRTPPacket *p) {
-  if(p->channel == CRTP_TUNNEL_CHANNEL_PING)
-    crtpTunnelPingHandler(p);
-  else if(p->channel == CRTP_TUNNEL_CHANNEL_PARAM)
-    crtpTunnelParametersHandler(p);
-  else if(p->channel == CRTP_TUNNEL_CHANNEL_COMMANDER)
-    crtpTunnelCommanderHandler(p);
-  else if(p->channel == CRTP_TUNNEL_CHANNEL_COMMAND) {
-    switch(p->data[0]) {
-      case CRTP_TUNNEL_COMMAND_TAKE_OFF:
-        tunnelSetDroneState(DRONE_STATE_FLYING);
-        break;
-      case CRTP_TUNNEL_COMMAND_LAND:
-        tunnelSetBehavior(TUNNEL_BEHAVIOR_LAND);
-        break;
-      case CRTP_TUNNEL_COMMAND_RTH:
-        //TODO
-        break;
-      case CRTP_TUNNEL_COMMAND_STOP:
-        tunnelSetDroneState(DRONE_STATE_IDLE);
-        break;
-    }
+    // Handle communication routines
+    tunnelCommUpdate();
   }
 }
 
 void tunnelInit() {
-  if(isInit)
-    return;
+  if(isInit) return;
 
   // State definitions
   tunnelSetDroneState(DRONE_STATE_IDLE);
@@ -141,15 +98,9 @@ void tunnelInit() {
   if(getDroneId() >= getNDrones())
     tunnelSetDroneState(DRONE_STATE_INACTIVE);
 
-  // Init submodules
-  tunnelPingInit();
-  tunnelParametersInit();
+  // Init submodules  
   tunnelCommanderInit();
-  tunnelSignalInit();
-  tunnelRelayInit();
-
-  // Subscribe to packet callbacks
-  crtpRegisterPortCB(CRTP_PORT_TUNNEL, crtpTunnelHandler);
+  tunnelCommInit();
 
   // Create the main tunnel task
   xTaskCreate(tunnelTask, TUNNELEXPLORER_TASK_NAME, TUNNELEXPLORER_TASK_STACKSIZE, NULL,
@@ -161,11 +112,8 @@ void tunnelInit() {
 bool tunnelTest() {
   bool pass = isInit;
 
-  pass &= tunnelPingTest();
-  pass &= tunnelParametersTest();
   pass &= tunnelCommanderTest();
-  pass &= tunnelSignalTest();
-  pass &= tunnelRelayTest();
+  pass &= tunnelCommTest();
 
   return pass;
 }
